Refactor PlaywrightTools into partial classes
